Add more checks for the paymentMethodsReponse - Add util function to do more checks in the paymentMethodsResponse prop. - Extract util functions to their own file. - Add tests.
